some_impovments

+++ TestTaskSumWebService/src/main/java/basavets/test/dto/response/ResponseDto.java
@@ -2,8 +2,9 @@ package basavets.test.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
+import java.math.BigInteger;
 
 @Builder(toBuilder = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record ResponseDto(String code, String description, Integer sum) {
+public record ResponseDto(String code, String description, BigInteger sum) {
 }
